+++ R/make-FV-profile.R
@@ -1,6 +1,7 @@
 #' Get Force-Velocity Profile
 #'
-#' Provides Force-Velocity (FV) profile suggested by Pierre Samozino and JB-Morin, et al.
+#' Provides Force-Velocity (FV) profile suggested by Pierre Samozino and JB-Morin, et al. (2016) and
+#'    Pierre Samozino and Nicolas Peyror, et al (2021).
 #' @inheritParams predict_kinematics
 #' @param RFmax_cutoff Time cut-off used to estimate \code{RFmax} and \code{Drf}. Default is 0.3s
 #' @return List containing the following elements:
@@ -17,6 +18,12 @@
 #'         \item{Drf}{Slope of Force Ratio (RF) and velocity. See References for more info}
 #'         \item{RSE_FV}{Residual standard error of the FV profile.}
 #'         \item{RSE_Drf}{Residual standard error of the RF-velocity profile}
+#'         \item{F0_poly}{Horizontal force when velocity=0, estimated using the analytics/polynomial method}
+#'         \item{F0_poly_rel}{\code{F0_poly} divided by \code{bodymass}}
+#'         \item{V0_poly}{Velocity when horizontal force=0, estimated using the analytics/polynomial method}
+#'         \item{Pmax_poly}{Maximal horizontal power, estimated using the analytics/polynomial method}
+#'         \item{Pmax_poly_rel}{\code{Pmax_poly} divided by \code{bodymass}}
+#'         \item{FV_slope_poly}{Slope of the FV profile, estimated using the analytics/polynomial method. See References for more info}
 #'         \item{data}{Data frame containing simulated data used to estimate parameters}
 #'     }
 #'
@@ -26,6 +33,11 @@
 #'     A simple method for measuring power, force, velocity properties, and mechanical effectiveness in
 #'     sprint running: Simple method to compute sprint mechanics. Scandinavian Journal of Medicine & Science
 #'     in Sports 26:648–658. DOI: 10.1111/sms.12490.
+#'
+#'     Samozino P, Peyrot N, Edouard P, Nagahara R, Jimenez‐Reyes P, Vanwanseele B, Morin J. 2022.
+#'      Optimal mechanical force‐velocity profile for sprint acceleration performance.
+#'       Scandinavian Journal of Medicine & Science in Sports 32:559–575. DOI: 10.1111/sms.14097.
+
 #' @examples
 #' data("jb_morin")
 #'
@@ -116,7 +128,20 @@ make_FV_profile <- function(MSS,
   V0 <- -F0_rel / Slope
 
   Pmax <- F0 * V0 / 4
-  Pmax_relative <- Pmax / bodymass
+  Pmax_rel <- Pmax / bodymass
+
+  ###########################
+  # Get polynomial solutions
+  # Get k value for the air resistance
+  k_rel <- get_air_resistance(velocity = 1, bodymass = bodymass, ...) / bodymass
+
+  tau <- MSS / MAC
+
+  F0_poly <- MAC * bodymass
+  V0_poly <- (1 / (2 * k_rel * tau)) * (1 - sqrt(1 - 4 * k_rel * tau * MSS))
+  Pmax_poly <- (F0_poly * V0_poly) / 4
+  Pmax_poly_rel <- Pmax_poly / bodymass
+  Slope_poly <- - (F0_poly / bodymass) / V0_poly
 
   # Return list
   new_fv_profile(
@@ -125,13 +150,19 @@ make_FV_profile <- function(MSS,
     F0_rel = F0_rel,
     V0 = V0,
     Pmax = Pmax,
-    Pmax_relative = Pmax_relative,
+    Pmax_rel = Pmax_rel,
     FV_slope = Slope,
     RFmax_cutoff = RFmax_cutoff,
     RFmax = RFmax,
     Drf = Drf,
     RSE_FV = summary(fv_profile)$sigma,
     RSE_Drf = summary(Drf_model)$sigma,
+    F0_poly = F0_poly,
+    F0_poly_rel = MAC,
+    V0_poly = V0_poly,
+    Pmax_poly = Pmax_poly,
+    Pmax_poly_rel = Pmax_poly_rel,
+    FV_slope_poly = Slope_poly,
     data = df
   )
 }